Adds hard coded token
The right way to do this is using the .env file or/and using a server
to make api calls. For the purpose of challenge and seeking keep
things simple, I created a read only token to access this public
informations. I believe in this case is not a securty issue.